refactor(auth): merge passkey challenge helper
