+++ tests/test_mfa_passkey_registration_errors.py
@@ -10,8 +10,8 @@ from app.helper.passkey import (
     PassKeyHelper,
     PassKeyRegistrationOriginMismatchError,
     PassKeyRegistrationVerificationError,
+    PasskeyChallengeStore,
 )
-from app.helper.passkey_challenge import PasskeyChallengeStore
 
 
 def _registration_request(user_id: int = 1) -> mfa_endpoint.PassKeyRegistrationFinish:
